WIP: hardware test partner round 1
Not sure which kicad files are necessary and which are local yet!

Goal: fixed "host" board (this board) with socket for _any_ Nucleo64 st
board, giving access to test:
* DAC->ADC (both directions)
* I2C (both directions)
* SPI (both directions)
* Uart (both directions)

a socket for a cheap fx2 based logic analyser will be included, so that
sigrok can be used to capture tests of the actual line states.