freertos-gadget0: properly use irq events
Lifted from the
https://www.freertos.org/deferred_interrupt_processing.html examples.

Problems uncovered: kernel irq split define was wildly wrong.

With this, we now actually have a proper _task_ that blocks until it has
things to do, rather than a gadget0 that only happened to work because
it was the only task running, and it was running at full speed.