freertos-spi-slave: usb controlled spi transfers
Not optimal, and not tested with a functional spi slave for the inbound
path, but functional valid usb controlled spi transfers in the out
direction.

Tested different bauds. cpha/cpol.... makes changes, but not sure it's
entirely correct?  Might be a sigrok problem?

Contains a "test" python script that makes a pair of transfers on each
run.  Based on the gadget0 tests